#4607 Translate captions in data dictionary depending on the user language setting
