+++ sormas-api/src/main/java/de/symeda/sormas/api/FacadeProvider.java
@@ -55,6 +55,7 @@ import de.symeda.sormas.api.hospitalization.HospitalizationFacade;
 import de.symeda.sormas.api.i18n.I18nFacade;
 import de.symeda.sormas.api.importexport.ExportFacade;
 import de.symeda.sormas.api.importexport.ImportFacade;
+import de.symeda.sormas.api.info.InfoFacade;
 import de.symeda.sormas.api.infrastructure.InfrastructureFacade;
 import de.symeda.sormas.api.infrastructure.PointOfEntryFacade;
 import de.symeda.sormas.api.infrastructure.PopulationDataFacade;
@@ -422,6 +423,10 @@ public class FacadeProvider {
 
 	public static CustomizableEnumFacade getCustomizableEnumFacade() {
 		return get().lookupEjbRemote(CustomizableEnumFacade.class);
+	}
+
+	public static InfoFacade getInfoFacae() {
+		return get().lookupEjbRemote(InfoFacade.class);
 	}
 
 	@SuppressWarnings("unchecked")
