+++ sormas-api/src/main/resources/enum.properties
@@ -1649,3 +1649,23 @@ LaboratoryDiagnosticEvidenceDetail.IMPRESSION_TEST=Impression test
 LaboratoryDiagnosticEvidenceDetail.WATER_SAMPLE=Water sample
 LaboratoryDiagnosticEvidenceDetail.OTHER=Other
 LaboratoryDiagnosticEvidenceDetail.PATHOGEN_FINE_TYPING_COMPLIANT_WITH_CASE=Pathogen fine typing compliant with the one of cases
+
+#EntityColumn
+EntityColumn.FIELD = Field
+EntityColumn.TYPE = Type
+EntityColumn.DATA_PROTECTION = Data protection
+EntityColumn.CAPTION = Caption
+EntityColumn.DESCRIPTION = Description
+EntityColumn.REQUIRED = Required
+EntityColumn.NEW_DISEASE = New disease
+EntityColumn.DISEASES = Diseases
+EntityColumn.OUTBREAKS = Outbreaks
+EntityColumn.IGNORED_COUNTRIES = Ignored countries
+EntityColumn.EXCLUSIVE_COUNTRIES = Exclusive countries
+
+#EnumColumn
+EnumColumn.TYPE = Type
+EnumColumn.VALUE = Value
+EnumColumn.CAPTION = Caption
+EnumColumn.DESCRIPTION = Description
+EnumColumn.SHORT = Short
