+++ sormas-api/src/main/resources/strings.properties
@@ -265,6 +265,7 @@ entityCommunities = Communities
 entityFacilities = Facilities
 entityPersonContactDetail = Person contact details
 entityUserRoles = User roles
+entityDataDictionary= Data dictionary
 
 # Error Messages
 errorAccessDenied=You do not have the required rights to view this page.
