+++ sormas-backend/src/test/java/de/symeda/sormas/backend/share/ExternalShareInfoFacadeEjbEjbTest.java
@@ -36,7 +36,7 @@ import de.symeda.sormas.api.utils.DateHelper;
 import de.symeda.sormas.backend.AbstractBeanTest;
 import de.symeda.sormas.backend.TestDataCreator;
 
-public class ExternalShareInfoFacadeEjbTest extends AbstractBeanTest {
+public class ExternalShareInfoFacadeEjbEjbTest extends AbstractBeanTest {
 
 	@Test
 	public void testGetIndexList() {
