+++ sormas-ui/src/main/java/de/symeda/sormas/ui/AboutView.java
@@ -17,21 +17,28 @@
  *******************************************************************************/
 package de.symeda.sormas.ui;
 
+import static de.symeda.sormas.ui.utils.DownloadUtil.createFileNameWithCurrentDate;
+
 import java.io.File;
+import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.nio.file.Files;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import org.apache.commons.io.IOUtils;
+import org.slf4j.LoggerFactory;
+
 import com.vaadin.icons.VaadinIcons;
 import com.vaadin.navigator.View;
 import com.vaadin.navigator.ViewChangeListener.ViewChangeEvent;
-import com.vaadin.server.ClassResource;
 import com.vaadin.server.ExternalResource;
 import com.vaadin.server.FileDownloader;
 import com.vaadin.server.FileResource;
+import com.vaadin.server.Page;
 import com.vaadin.server.StreamResource;
 import com.vaadin.server.VaadinService;
 import com.vaadin.server.VaadinServletRequest;
@@ -42,6 +49,7 @@ import com.vaadin.ui.Button;
 import com.vaadin.ui.CustomLayout;
 import com.vaadin.ui.Label;
 import com.vaadin.ui.Link;
+import com.vaadin.ui.Notification;
 import com.vaadin.ui.VerticalLayout;
 import com.vaadin.ui.themes.ValoTheme;
 
@@ -51,10 +59,12 @@ import de.symeda.sormas.api.caze.classification.ClassificationHtmlRenderer;
 import de.symeda.sormas.api.feature.FeatureType;
 import de.symeda.sormas.api.i18n.Captions;
 import de.symeda.sormas.api.i18n.I18nProperties;
+import de.symeda.sormas.api.i18n.Strings;
 import de.symeda.sormas.api.utils.InfoProvider;
 import de.symeda.sormas.ui.utils.ButtonHelper;
 import de.symeda.sormas.ui.utils.CssStyles;
 import de.symeda.sormas.ui.utils.DownloadUtil;
+import de.symeda.sormas.ui.utils.ExportEntityName;
 
 @SuppressWarnings("serial")
 public class AboutView extends VerticalLayout implements View {
@@ -157,8 +167,21 @@ public class AboutView extends VerticalLayout implements View {
 				Button dataDictionaryButton =
 					ButtonHelper.createButton(Captions.aboutDataDictionary, null, ValoTheme.BUTTON_LINK, CssStyles.BUTTON_COMPACT);
 				documentsLayout.addComponent(dataDictionaryButton);
-				FileDownloader dataDictionaryDownloader = new FileDownloader(new ClassResource("/doc/SORMAS_Data_Dictionary.xlsx"));
-				dataDictionaryDownloader.extend(dataDictionaryButton);
+
+				new FileDownloader(new StreamResource(() -> new DownloadUtil.DelayedInputStream((out) -> {
+					try {
+						String documentPath = FacadeProvider.getInfoFacae().generateDataDictionary();
+						IOUtils.copy(Files.newInputStream(new File(documentPath).toPath()), out);
+					} catch (IOException e) {
+						LoggerFactory.getLogger(DownloadUtil.class).error(e.getMessage(), e);
+						new Notification(
+							I18nProperties.getString(Strings.headingExportUserRightsFailed),
+							I18nProperties.getString(Strings.messageUserRightsExportFailed),
+							Notification.Type.ERROR_MESSAGE,
+							false).show(Page.getCurrent());
+					}
+				}, (e) -> {
+				}), createFileNameWithCurrentDate(ExportEntityName.DATA_DICTIONARY, ".xlsx"))).extend(dataDictionaryButton);
 			}
 
 			// This link is hidden until an updated version of the document is provided
